requirements.txt
absl-py==1.0.0
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
click==8.1.3
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
distlib==0.3.5
distro==1.7.0
executing==1.1.1
filelock==3.7.1
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h11==0.13.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
ipython==7.34.0
jedi==0.18.1
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.6
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
poppler-utils==0.1.0
prompt-toolkit==3.0.31
protobuf==3.19.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
selenium==4.4.3
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.1
tabula==1.0.5
tabula-py==2.4.0
tabulate==0.9.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.64.0
traitlets==5.4.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.16.2
wcwidth==0.2.5
webdriver-manager==3.8.3
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.2.0
zipp==3.6.0
